requirements_test.txt
pytest-homeassistant-custom-component

# Additional test utilities
pytest-cov
pytest-sugar
python-dotenv

# Linting & formatting (match pre-commit versions)
ruff

# Component dependencies
# (no external dependencies - API client is built-in)
